[not verified] Convert JetpackFooter stories to TypeScript
